modifs

+++ Appliweb1/src/Pack/Facade.java
@@ -3,14 +3,14 @@ package Pack;
 import java.util.ArrayList;
 import java.util.List;
 
+import javax.ejb.Singleton;
 import javax.persistence.*;
 
+@Singleton
 public class Facade implements Facade_Itf {
 
 	@PersistenceContext 
 	private EntityManager em;
-
-	public Facade() {};
 
 	// Utilisé pour l'inscription d'un utilisateur
 	public void addUser(String login, String password) {
@@ -29,6 +29,12 @@ public class Facade implements Facade_Itf {
 		return user.getPassword().equals(password);
 	}
 
+	// Utilisé pour récupérer un utilisateur
+	public Client getClient(String login) {
+		Client user = em.find(Client.class, login);
+		return user;
+	}
+
 	// Utilisé pour récupérer toutes les musiques
 	public List<List<String>> getAllMusic() {
 		List<List<String>> list = new ArrayList<List<String>>();
@@ -45,29 +51,35 @@ public class Facade implements Facade_Itf {
 	}
 
 	// Utilisé pour récupérer les musiques favorites d'un utilisateur
-	public List<Integer> getFavs(Integer idUser) {
-		List<Integer> list = em.createQuery("SELECT m.id FROM Music m JOIN m.fav f WHERE f.client.id = :idUser").getResultList();
+	public List<Favorite> getFavs(String login) {
+		Client user = em.find(Client.class, login);
+		List<Favorite> list = user.getFavorites();
 		return list;
 	}
 
 	// Utilisé pour ajouter des musiques favorites à un utilisateur
-	public void addFavs(Integer idUser, List<Integer> idMusic) {
-		Client user = em.find(Client.class, idUser);
+	public void addFavs(String login, List<Integer> idMusic) {
+		Client user = em.find(Client.class, login);
+		List<Favorite> list = user.getFavorites();
 		for (Integer id : idMusic) {
 			Music music = em.find(Music.class, id);
 			Favorite fav = new Favorite();
 			fav.setClient(user);
 			fav.setMusic(music);
+			list.add(fav);
 			em.persist(fav);
 		}
 	}
 
 	// Utilisé pour supprimer des musiques favorites à un utilisateur
-	public void delFavs(Integer idUser, List<Integer> idMusic) {
-		for (Integer id : idMusic) {
-			List<Favorite> list = em.createQuery("SELECT f FROM Favorite f WHERE f.client.id = :idUser AND f.music.id = :idMusic").getResultList();
-			for (Favorite fav : list) {
+	public void delFavs(String login, List<Integer> idMusic) {
+		Client user = em.find(Client.class, login);
+		List<Favorite> list = user.getFavorites();
+		for (Favorite fav : list) {
+			int id = fav.getMusic().getId();
+			if (idMusic.contains(id)) {
 				em.remove(fav);
+				list.remove(fav);
 			}
 		}
 	}
